requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
python-multipart==0.0.12
pydantic-settings==2.6.0
pyyaml==6.0.2
SpeechRecognition==3.11.0
pyaudio==0.2.14
transformers==4.45.2
torch==2.5.0+cpu
google-generativeai==0.6.10